SAMPY/requirements.txt
cherrypy
jinja2
pyyaml
